pyproject.toml
[tool.poetry]
name = "scooter-kg"
version = "0.1.0"
description = ""
authors = ["muratbek2255 <muratbek2255@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
django-filter = "^22.1"
drf-yasg = "^1.21.3"
django-environ = "^0.9.0"
celery = "^5.2.7"
redis = "^4.3.4"
social-auth-app-django = "^5.0.0"
django-rest-framework-social-oauth2 = "^1.1.0"
qrcode = "^7.3.1"
django-allauth = "^0.51.0"
requests = "^2.28.1"
django-crispy-forms = "^1.14.0"
twilio = "^7.12.0"
pyotp = "^2.6.0"
python-decouple = "^3.6"
django-modeltranslation = "^0.18.4"
firebase-admin = "^5.2.0"
django-redis = "^5.2.0"
utils = "^1.0.1"
Faker = "^13.15.1"
stripe = "^4.0.2"
django-celery-beat = "^2.3.0"
six = "^1.16.0"
django-jet-4 = "^0.0.1"
fcm-django = "^1.0.12"
pyfcm = "^1.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
